cleanup, restructure to pyproject.toml, verioningit
