--- .codecov.yml
@@ -1,14 +1,0 @@
-# Codecov configuration to make it a bit less noisy
-coverage:
-  status:
-    patch: false
-    project:
-      default:
-        threshold: 50%
-comment:
-  layout: "header"
-  require_changes: false
-  branches: null
-  behavior: default
-  flags: null
-  paths: null
