--- .lgtm.yml
@@ -1,12 +1,0 @@
-# Configure LGTM for this package
-
-extraction:
-  python:  # Configure Python
-    python_setup:  # Configure the setup
-      version: 3  # Specify Version 3
-path_classifiers:
-  library:
-    - versioneer.py  # Set Versioneer.py to an external "library" (3rd party code)
-    - devtools/*
-  generated:
-    - responsefun/_version.py
