--- .travis.yml
@@ -1,45 +1,0 @@
-language: python
-
-# Run jobs on container-based infrastructure, can be overridden per job
-
-matrix:
-  include:
-    # Extra includes for OSX since python language is not available by default on OSX
-    - os: osx
-      language: generic
-      env: PYTHON_VER=3.6
-
-
-    # Pip can use Travis build-in Python
-    - os: linux
-      python: 3.6
-    - os: linux
-      dist: xenial  # Travis Trusty image does not have Python 3.7, Xenial does
-      python: 3.7
-
-
-before_install:
-    # Additional info about the build
-  - uname -a
-  - df -h
-  - ulimit -a
-
-    # Install the Python environment
-  - source devtools/travis-ci/before_install.sh
-  - python -V
-
-install:
-
-    # Install the package locally
-  - pip install -U pytest pytest-cov codecov
-  - pip install -e .
-
-
-script:
-  - pytest -v --cov=responsefun responsefun/tests/
-
-notifications:
-    email: false
-
-after_success:
-  - codecov
