+++ README.md
@@ -1,11 +1,11 @@
 ResponseFun
 ==============================
 
-Fun with Response Functions
+Fun with Response Functions in the Algebraic Diagrammatic Construction Framework.
 
 ### Copyright
 
-Copyright (c) 2023, The responsefun Developers
+Copyright (c) 2023, The `responsefun` Developers
 
 
 #### Acknowledgements
