--- devtools/conda-envs/test_env.yaml
@@ -1,14 +1,0 @@
-name: test
-channels:
-dependencies:
-    # Base depends
-  - python
-  - pip
-
-    # Testing
-  - pytest
-  - pytest-cov
-
-    # Pip-only installs
-  - pip:
-    - codecov
