--- devtools/conda-recipe/meta.yaml
@@ -1,27 +1,0 @@
-package:
-  name: responsefun
-  version: "0.0.0"
-
-source:
-  path: ../..
-
-build:
-  number: 0
-
-requirements:
-  host:
-    - python
-    - setuptools
-
-  run:
-    - python
-
-test:
-  requires:
-    - pytest
-  imports:
-    - responsefun
-
-about:
-  home: add_url_here
-  license: LGPLv3 License
