+++ examples/alpha_sos_expr.py
@@ -3,17 +3,21 @@ Specify the SOS expression of a response function symbolically using the example
 """
 
 from sympy import symbols
-from responsefun.response_operators import (
-    DipoleOperator,
+from responsefun.ResponseOperator import (
     TransitionFrequency
 )
-from responsefun.sum_over_states import TransitionMoment
+from responsefun.symbols_and_labels import (
+    O, f, gamma, n, m, p, k,
+    op_a, op_b, op_c, op_d,
+    w_f, w_n, w_m, w_p, w_k, w, w_o, w_1, w_2, w_3,
+)
+from responsefun.SumOverStates import TransitionMoment
 
 # define symbols and operators
-O, n, gamma, w = symbols(r"0, n, \gamma, w", real=True)
-op_a = DipoleOperator("A")
-op_b = DipoleOperator("B")
-w_n = TransitionFrequency("n", real=True)
+# O, n, gamma, w = symbols(r"0, n, \gamma, w", real=True)
+# op_a = DipoleOperator("A")
+# op_b = DipoleOperator("B")
+# w_n = TransitionFrequency("n", real=True)
 
 # define symbolic SOS expression
 alpha_sos_expr = (
