+++ responsefun/__init__.py
@@ -7,8 +7,4 @@ Fun with Response Functions
 from responsefun import *
 
 # Handle versioneer
-from ._version import get_versions
-versions = get_versions()
-__version__ = versions['version']
-__git_revision__ = versions['full-revisionid']
-del get_versions, versions
+from ._version import __version__
