+++ setup.cfg
@@ -8,20 +8,3 @@ omit =
     # Omit generated versioneer
     responsefun/_version.py
 
-[yapf]
-# YAPF, in .style.yapf files this shows up as "[style]" header
-COLUMN_LIMIT = 119
-INDENT_WIDTH = 4
-USE_TABS = False
-
-[flake8]
-# Flake8, PyFlakes, etc
-max-line-length = 119
-
-[versioneer]
-# Automatic version numbering scheme
-VCS = git
-style = pep440
-versionfile_source = responsefun/_version.py
-versionfile_build = responsefun/_version.py
-tag_prefix = ''
