--- setup.py
@@ -1,52 +1,0 @@
-"""
-ResponseFun
-Fun with Response Functions
-"""
-from setuptools import setup
-import versioneer
-
-short_description = __doc__.split("\n")
-
-try:
-    with open("README.md", "r") as handle:
-        long_description = handle.read()
-except:
-    long_description = "\n".join(short_description[2:]),
-
-
-setup(
-    # Self-descriptive entries which should always be present
-    name='responsefun',
-    author='Maximilian Scheurer',
-    author_email='info@maxscheurer.com',
-    description=short_description[0],
-    long_description=long_description,
-    long_description_content_type="text/markdown",
-    version=versioneer.get_version(),
-    cmdclass=versioneer.get_cmdclass(),
-    license='LGPLv3',
-
-    # Which Python importable modules should be included when your package is installed
-    packages=['responsefun'], 
-
-    # Optional include package data to ship with your package
-    # Comment out this line to prevent the files from being packaged with your software
-    # Extend/modify the list to include/exclude other items as need be
-    install_requires=['sympy'],
-    package_data={'responsefun': ["data/*.dat"]
-                  }
-
-    # Additional entries you may want simply uncomment the lines you want and fill in the data
-    # author_email='me@place.org',      # Author email
-    # url='http://www.my_package.com',  # Website
-    # install_requires=[],              # Required packages, pulls from pip if needed; do not use for Conda deployment
-    # platforms=['Linux',
-    #            'Mac OS-X',
-    #            'Unix',
-    #            'Windows'],            # Valid platforms your code works on, adjust to your flavor
-    # python_requires=">=3.5",          # Python version restrictions
-
-    # Manual control if final package is compressible or not, set False to prevent the .egg from being made
-    # zip_safe=False,
-
-)
